Moved base entities from sprites to entities folder https://github.com/achobanov/Web-Game/issues/11
