+++ Source/Client/source/events/add-entity-event.ts
@@ -1,5 +1,5 @@
 import { IEvent } from "../services/events-service";
-import { IEntity } from "../objects/sprites/entity";
+import { IEntity } from "../objects/entities/entity";
 
 export default class AddEntityEvent implements IEvent {
     static Key = 'add-entity';
