+++ Source/Client/source/objects/entities/entity.ts
