+++ Source/Client/source/objects/entities/sprite.ts
